requirements.txt
# Core
torch>=2.2.0
transformers>=4.45.0
accelerate>=0.30.0
datasets>=2.19.0
evaluate>=0.4.1
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.26.0

# ROUGE metric
rouge-score>=0.1.2

# GPTQ (Owner: Sakshi and Mahi)
auto-gptq>=0.7.1

# AWQ (Owner: Subrat, Anurag)
autoawq>=0.2.5

# Optional: faster inference
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
protobuf>=4.25.0
